# swagger_latest.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AITasker Modular API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Projects/proposal/{proposalId}": {
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Proposals/submit-proposal": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "JobPostId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "ExpertId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "BidAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "EstimatedDuration": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Introduction": {
                    "type": "string"
                  },
                  "Implementation": {
                    "type": "string"
                  },
                  "Portfolio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "PortfolioUrl": {
                    "type": "string"
                  },
                  "Attachment": {
                    "type": "string",
                    "format": "binary"
                  },
                  "AttachmentUrl": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "JobPostId": {
                  "style": "form"
                },
                "ExpertId": {
                  "style": "form"
                },
                "BidAmount": {
                  "style": "form"
                },
                "EstimatedDuration": {
                  "style": "form"
                },
                "Introduction": {
                  "style": "form"
                },
                "Implementation": {
                  "style": "form"
                },
                "Portfolio": {
                  "style": "form"
                },
                "PortfolioUrl": {
                  "style": "form"
                },
                "Attachment": {
                  "style": "form"
                },
                "AttachmentUrl": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/job/{jobPostId}": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "jobPostId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/expert/{expertId}": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "expertId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/{id}/status": {
      "put": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptProposalStatusDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptProposalStatusDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptProposalStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/{id}": {
      "put": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "BidAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "EstimatedDuration": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Introduction": {
                    "type": "string"
                  },
                  "Implementation": {
                    "type": "string"
                  },
                  "Portfolio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "PortfolioUrl": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "BidAmount": {
                  "style": "form"
                },
                "EstimatedDuration": {
                  "style": "form"
                },
                "Introduction": {
                  "style": "form"
                },
                "Implementation": {
                  "style": "form"
                },
                "Portfolio": {
                  "style": "form"
                },
                "PortfolioUrl": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/{id}/generate-milestone-md": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "taskCount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deadlineDays",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/analyze-job-to-usecases/{jobPostId}": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "jobPostId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/expert-ai-chat-session": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpertAiChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpertAiChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpertAiChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proposals/expert-ai-chat-history": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "name": "jobPostId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expertId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcceptProposalStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExpertAiChatRequest": {
        "type": "object",
        "properties": {
          "jobPostId": {
            "type": "string",
            "format": "uuid"
          },
          "expertId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token below.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}